+++ pytorch_lightning/trainer/trainer.py
@@ -45,7 +45,7 @@ from pytorch_lightning.plugins import (
     ApexMixedPrecisionPlugin,
     DDPSpawnStrategy,
     NativeMixedPrecisionPlugin,
-    ParallelPlugin,
+    ParallelStrategy,
     PLUGIN_INPUT,
     PrecisionPlugin,
     Strategy,
@@ -1927,7 +1927,7 @@ class Trainer(
 
     @property
     def distributed_sampler_kwargs(self) -> Optional[dict]:
-        if isinstance(self.training_type_plugin, ParallelPlugin):
+        if isinstance(self.training_type_plugin, ParallelStrategy):
             return self.training_type_plugin.distributed_sampler_kwargs
 
     @property