+++ tests/accelerators/test_accelerator_connector.py
@@ -32,7 +32,7 @@ from pytorch_lightning.plugins import (
     DDPSpawnStrategy,
     DDPStrategy,
     DeepSpeedStrategy,
-    ParallelPlugin,
+    ParallelStrategy,
     PrecisionPlugin,
     SingleDeviceStrategy,
 )
@@ -427,7 +427,7 @@ def test_plugin_accelerator_choice(accelerator: Optional[str], plugin: str):
 @mock.patch("torch.cuda.device_count", return_value=2)
 @pytest.mark.parametrize("gpus", [1, 2])
 def test_accelerator_choice_multi_node_gpu(
-    mock_is_available, mock_device_count, tmpdir, accelerator: str, plugin: ParallelPlugin, gpus: int
+    mock_is_available, mock_device_count, tmpdir, accelerator: str, plugin: ParallelStrategy, gpus: int
 ):
     with pytest.deprecated_call(match=r"accelerator=.*\)` has been deprecated"):
         trainer = Trainer(accelerator=accelerator, default_root_dir=tmpdir, num_nodes=2, gpus=gpus)